requirements.txt
absl-py==2.0.0
aiohttp==3.9.0
aiosignal==1.3.1
alembic==1.12.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
colorlog==6.7.0
contourpy==1.2.0
cycler==0.12.1
datasets==2.15.0
decorator==4.4.2
dill==0.3.7
docopt==0.6.2
editdistance==0.6.2
efficientnet==1.0.0
einops==0.7.0
essential-generators==1.0
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
google-api-core==2.14.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-cloud-speech==2.22.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.12.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
huggingface-hub==0.19.4
huggingsound==0.1.6
hyperframe==6.0.1
HyperPyYAML==1.2.2
idna==3.5
imageio==2.33.0
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jiwer==2.6.0
joblib==1.3.2
julius==0.2.7
keras==2.13.1
Keras-Applications==1.0.8
keras-ocr @ git+https://github.com/faustomorales/keras-ocr.git@e8d34a46f07d50158e1d86d6c617e99bfe99e2f8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
lightning==2.1.2
lightning-utilities==0.10.0
llvmlite==0.41.1
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.0
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
noisereduce==3.0.0
norbert==0.2.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cudnn-cu11==8.9.5.29
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.8.1.78
opencv-python==4.5.1.48
opt-einsum==3.3.0
optuna==3.4.0
packaging==23.2
pandas==1.5.3
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
primePy==1.3
proglog==0.1.10
proto-plus==1.22.3
protobuf==4.25.1
pyannote.audio==3.1.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydub==0.25.1
pyee==11.1.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-ffmpeg==2.0.4
python-version==0.0.2
pytorch-lightning==2.1.2
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
pyvad==0.2.0
PyYAML==6.0.1
rapidfuzz==2.13.7
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rich==13.7.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
semver==3.0.2
sentencepiece==0.1.99
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.23
srt==3.5.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.15.0
torch==2.1.1+cu118
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.1
torchmetrics==1.2.0
torchvision==0.16.1+cu118
tqdm==4.66.1
transformers==4.35.2
typer==0.3.2
typing_extensions==4.5.0
urllib3==2.1.0
validators==0.22.0
vosk==0.3.45
wav2lip==1.2.4
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
